requirements.txt
aiowebthing
snownlp
